HIVE-21480: Fixed flaky and broken test TestHiveMetaStore.testJDOPersistenceManagerCleanup
This test was not correctly counting the number of
objects in the PersistenceManager cache before and after HiveMetaStoreClient.close(). The
getJDOPersistenceManagerCacheSize() internal helper method did not use
the updated fields present in the metastore classes, and was
consistently returning -1. Additionally, there was a chance to cause
flakiness since the object count before and after close() could
differ depending on lingering objects from previous
tests or setup.

Modified the helper method to use the new
fields, and fixed the flakiness on this test.